feat: command line options in test-rt to filter the tests to run

+++ c/yarray.h
@@ -1,18 +1,20 @@
 #ifndef _Y_ARRAY_H_
 #define _Y_ARRAY_H_
 
+#include <stdint.h>
+
 typedef struct {
-    unsigned long long len;
+    uint64_t len;
     void * data;
 } _yrt_array_;
 
 typedef struct {
-    unsigned long long len;
+    uint64_t len;
     char * data;
 } _yrt_c8_array_;
 
 typedef struct {
-    unsigned long long len;
+    uint64_t len;
     unsigned int * data;
 } _yrt_c32_array_;
